PPOX New File Organization
PPOX save files have miraculously taken on attributes!! Greatly
shortening the file size and making everything look overall much nicer.

Also, PPOX recognizes if you are trying to load in a file that uses a
different version from the one the editor is using, and yells at you for
trying to do so.